Remove diffraction_tools.py, diffsims + signal.dpc takes over
